+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 - Implementing WebSocket
 - Change the header class name: `fa fa-skyatlas` font size.
 - Change in the local storage names.
-- Bi-language (English/French[70% done])
+- Bi-language (English/French[70% done]).
+- Auto-change to last selected language.
+- Equalization fr-en json.
+- Adding notification on upload status.
 
 ### 0.3.0
 - Add weather setting.